+++ src/light/antagonistColorLightness.ts
@@ -1,9 +1,11 @@
-import type { ContrastConfig_WTF } from "../contrast/contrastConfig";
+import type { ContrastConfig_PREPARED } from "../contrast/contrastConfig";
 
 import { convertToOklch_orThrow } from "../utils/culoriUtils";
 import { log } from "../utils/log";
 
-export function antagonistColorLightness(contrastConfig: ContrastConfig_WTF) {
+export function antagonistColorLightness(
+  contrastConfig: ContrastConfig_PREPARED
+) {
   let oklch = convertToOklch_orThrow(contrastConfig.colorAntagonist);
   log("culori > convertToOklch /// antagonistColorLightness");
   return oklch.l;
